[#184] finally got rid of neofs.py
