+++ robot/testsuites/integration/object/object_complex.robot
@@ -5,7 +5,6 @@ Variables   common.py
 Library     container.py
 Library     complex_object_actions.py
 Library     neofs_verbs.py
-Library     neofs.py
 Library     storage_policy.py
 Library     utility_keywords.py
