+++ robot/testsuites/integration/object/storage_group/sg_of_complex_objects.robot
@@ -1,7 +1,6 @@
 *** Settings ***
 Variables   common.py
 
-Library     neofs.py
 Library     container.py
 Library     neofs_verbs.py
 Library     storage_group.py
