+++ robot/testsuites/integration/services/http_gate.robot
@@ -3,7 +3,6 @@ Variables   common.py
 Variables   wellknown_acl.py
 
 Library     container.py
-Library     neofs.py
 Library     neofs_verbs.py
 Library     http_gate.py
 Library     storage_policy.py
@@ -25,10 +24,10 @@ NeoFS HTTP Gateway
     [Setup]             Setup
                         Make Up    ${INCLUDE_SVC}
 
-    ${WALLET}   ${_}     ${_} =   Prepare Wallet And Deposit
-    ${CID} =    Create container                    ${WALLET}    rule=${PLACEMENT_RULE}  basic_acl=${PUBLIC_ACL}
-    ${FILE}    ${HASH} =    Generate file    ${SIMPLE_OBJ_SIZE}
-    ${FILE_L}    ${L_HASH} =    Generate file    ${COMPLEX_OBJ_SIZE}
+    ${WALLET}   ${_}     ${_} =     Prepare Wallet And Deposit
+    ${CID} =                        Create container                    ${WALLET}    rule=${PLACEMENT_RULE}  basic_acl=${PUBLIC_ACL}
+    ${FILE}    ${HASH} =            Generate file    ${SIMPLE_OBJ_SIZE}
+    ${FILE_L}    ${L_HASH} =        Generate file    ${COMPLEX_OBJ_SIZE}
 
     ${S_OID} =          Put object                 ${WALLET}    ${FILE}      ${CID}
     ${L_OID} =          Put object                 ${WALLET}    ${FILE_L}    ${CID}
@@ -38,8 +37,8 @@ NeoFS HTTP Gateway
     @{GET_NODE_LIST} =  Get nodes without object            ${WALLET}    ${CID}    ${S_OID}
     ${NODE} =           Evaluate                            random.choice($GET_NODE_LIST)    random
 
-    ${GET_OBJ_S} =      Get object               ${WALLET}     ${CID}    ${S_OID}    ${EMPTY}    s_file_read    ${NODE}
-    ${FILEPATH} =       Get via HTTP Gate                   ${CID}    ${S_OID}
+    ${GET_OBJ_S} =      Get object               ${WALLET}      ${CID}    ${S_OID}    ${EMPTY}    s_file_read    ${NODE}
+    ${FILEPATH} =       Get via HTTP Gate        ${CID}         ${S_OID}
 
     ${PLAIN_FILE_HASH} =    Get file hash       ${GET_OBJ_S}
     ${GATE_FILE_HASH} =     Get file hash       ${FILEPATH}
@@ -49,8 +48,8 @@ NeoFS HTTP Gateway
     @{GET_NODE_LIST} =  Get nodes without object            ${WALLET}    ${CID}    ${L_OID}
     ${NODE} =           Evaluate                            random.choice($GET_NODE_LIST)    random
 
-    ${GET_OBJ_L} =      Get object               ${WALLET}     ${CID}    ${L_OID}    ${EMPTY}    l_file_read    ${NODE}
-    ${FILEPATH} =       Get via HTTP Gate                   ${CID}    ${L_OID}
+    ${GET_OBJ_L} =      Get object               ${WALLET}      ${CID}    ${L_OID}    ${EMPTY}    l_file_read    ${NODE}
+    ${FILEPATH} =       Get via HTTP Gate        ${CID}         ${L_OID}
 
     ${PLAIN_FILE_HASH} =    Get file hash       ${GET_OBJ_L}
     ${GATE_FILE_HASH} =     Get file hash       ${FILEPATH}
