+++ robot/testsuites/integration/services/s3_gate_object.robot
@@ -5,7 +5,6 @@ Library     Collections
 Library     OperatingSystem
 
 Library     container.py
-Library     neofs.py
 Library     s3_gate.py
 Library     utility_keywords.py
